final
The developed project is a Solar System Explorer that provides a detailed and dynamic visualisation of the orbital configuration around the Sun. It has functions such as zoom, camera movement and navigation between planets, which facilitates the study of the solar system. The representation uses real values to ensure visualisation accuracy.